set component folder and fix index
