+++ frontend/src/Game/components/MatchList/MatchList.tsx
@@ -1,7 +1,7 @@
 import React, { type ReactElement, useEffect, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
-import { getMatches } from '../../utils/matchAxios'
-import { getAllUsersRequest } from '../../utils/userAxios'
+import { getMatches } from '../../../utils/matchAxios'
+import { getAllUsersRequest } from '../../../utils/userAxios'
 import ListGroup from 'react-bootstrap/ListGroup'
 import axios from 'axios'
