+++ frontend/src/Game/index.tsx
@@ -1,4 +1,4 @@
-import { Game } from './components/Game'
-import { MatchList } from './components/MatchList'
+import { Game } from './components/Game/Game'
+import { MatchList } from './components/MatchList/MatchList'
 
 export { Game, MatchList }
